Update to support 1.5
Stacking/splitting confirmed to work as expected. Recipes that produce multiple meals are either all poisoned or all not poisoned, consistent with vanilla behavior. In playtesting I found it surprisingly difficult to trigger food poisoning due to filthyKitchen, even with a room that was designated as very dirty. In such a room this should be happening 5% of the time, but in my test I only saw this about 2% of the time. Could be a matter of luck in my case, but worth keeping an eye on.